+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.2 (unreleased)
 ----------------
 
+- Added task content interface to differentiate from behavior
+  [sgeulette]
 - Added assigned_group and due_date indexes. Added indexer methods avoiding acquisition for children.
   [sgeulette]
 - Changed workflow to use 'Request review' guard permission to differentiate Reviewer and editor transitions (like in plone workflows)